DAQ time difference
Get data from MINERvA and MINOS data acquisition system. The MINOS is getting its time information from GPS whereas MINERvA is getting it time information from a different time server. This created few failures when there are time differences between the detectors. Which then resulted as a failure to reconstruct a particle track from MINERvA to MINOS. The timings for both the DAQ are set from the same time server. The scripts spit the time difference between both DAQs, plots it and send it to the dynamic shift website.